requirements.txt
google-cloud-logging==3.11.0
flask==3.0.3
gunicorn==22.0.0
pytest==8.2.0
